+++ EvtGenExternal/EvtVinciaQED.hh
@@ -18,45 +18,57 @@
 * along with EvtGen.  If not, see <https://www.gnu.org/licenses/>.     *
 ***********************************************************************/
 
-#ifndef EVTEXTERNALGENLIST_HH
-#define EVTEXTERNALGENLIST_HH
+#ifdef EVTGEN_VINCIA
+#ifndef EVTVINCIAQED_HH
+#define EVTVINCIAQED_HH
 
 #include "EvtGenBase/EvtAbsRadCorr.hh"
-#include "EvtGenBase/EvtDecayBase.hh"
+#include "EvtGenBase/EvtId.hh"
+#include "EvtGenBase/EvtVector4R.hh"
 
-#include <list>
+#include "EvtGenExternal/EvtPythiaRandom.hh"
 
-// Description: A factory type method to create engines for external physics
-// generators like Pythia.
+#include "Pythia8/ParticleData.h"
+#include "Pythia8/Pythia.h"
+#include "Pythia8/ShowerModel.h"
 
-class EvtExternalGenList {
+#include <string>
+
+class EvtParticle;
+
+/*
+ * Description: EvtGen's module for generation of
+ *              QED final state radiation using VinciaQED.
+ */
+
+class EvtVinciaQED : public EvtAbsRadCorr {
   public:
-    EvtExternalGenList( bool convertPythiaCodes = false,
-                        std::string pythiaXmlDir = "",
-                        std::string photonType = "gamma",
-                        bool useEvtGenRandom = true,
-                        bool seedTauolaFortran = true,
-                        bool useTauolaRadiation = false,
-                        double infraredCutOffTauola = 1.0e-7 );
+    EvtVinciaQED( std::string xmlDir = "./xmldoc", bool useEvtGenRandom = true,
+                  double infraredCutOff = 1.0e-7 );
 
-    virtual ~EvtExternalGenList();
+    void initialise() override;
 
-    std::list<EvtDecayBase*> getListOfModels();
-
-    EvtAbsRadCorr* getPhotosModel( const double infraredCutOff = 1.0e-7,
-                                   const double maxWtInterference = 64.0 );
-
-    EvtAbsRadCorr* getSherpaPhotonsModel( const double infraredCutOff = 1.0e-7,
-                                          const int mode = 2,
-                                          const int useME = 0 );
-
-    EvtAbsRadCorr* getVinciaQEDModel( const double infraredCutOff = 1.0e-7 );
+    void doRadCorr( EvtParticle* theParticle ) override;
 
   private:
-    std::string m_photonType;
-    std::string m_pythiaXmlDir;
+    void storeDaughterInfo( Pythia8::Event& theEvent, EvtParticle* theParticle,
+                            int startInt, int ancestorId );
 
-    bool m_useEvtGenRandom;
+    // Specify if we are going to use the random number generator (engine) from EvtGen for Pythia 8.
+    bool m_useEvtGenRandom{ true };
+
+    // Minimum photon Energy (infrared cut-off).
+    double m_infraredCutOff{ 1.0e-7 };
+
+    EvtId m_gammaId = EvtId( -1, -1 );
+    long int m_gammaPDG{ 22 };
+
+    std::unique_ptr<Pythia8::Pythia> m_vinciaPythiaGen;
+    std::shared_ptr<EvtPythiaRandom> m_evtgenRandom;
+
+    bool m_initialised{ false };
 };
+
+#endif
 
 #endif